backups/BACKUPS_old/SLH_Project_20260316_2329/requirements.txt
fastapi
uvicorn
python-dotenv
redis
aiogram
asyncpg
SQLAlchemy
psycopg2-binary
